+++ apps/openmw/mwgui/savegamedialog.cpp
@@ -548,7 +548,7 @@ namespace MWGui
         }
         else if (arg.button == SDL_CONTROLLER_BUTTON_Y)
         {
-            uint32_t index = mCharacterSelection->getIndexSelected();
+            size_t index = mCharacterSelection->getIndexSelected();
             index = wrap(index + 1, mCharacterSelection->getItemCount());
             mCharacterSelection->setIndexSelected(index);
             onCharacterSelected(mCharacterSelection, index);
